add a header for sk-libfido2.c
